// packages/json-joy/src/json-patch/codec/compact/types.ts
export type OPCODE_SPLIT = OPCODE.split | 'split';
// ...
export type CompactSplitOp =
  | [opcode: OPCODE_SPLIT, path: string | Path, pos: number]
  | [opcode: OPCODE_SPLIT, path: string | Path, pos: number, props?: object | null];